# AIRFLOW-128 Optimize and refactor process_dag