@bolkedebruin 2016-05-23
Builds should be fine, just cant restart them: https://travis-ci.org/bolkedebruin/airflow.