@criccomini 2016-05-31
LGTM

Few questions/nits. Also, there are a lot of TODOs. I'm assuming these will get cleaned up as you continue to merge more and more of your uber-patch?

Also^2, should get a +1 from someone other than me, as well.