@jlowin 2016-05-31
Looking great, Bolke. However I think it might introduce one new corner case:

1. TI is set as "SCHEDULED" and sent to the executor
2. executor has a failure loading the TI (so the TI never gets set to FAILED or even RUNNING)
3. I think that TI is orphaned forever? Schedulers will never touch it again because it looks like it has been SCHEDULED. 

Perhaps some sort of garbage collection is needed for zombie/orphaned tasks?