@bolkedebruin 2016-05-31
@jlowin as discussed on Gitter your corner case is correct, however it should be addressed in a separate PR as fixing that corner case will require further fundamental changes (probably LocalTaskJob and Executor areas).