@bolkedebruin 2016-06-01
@criccomini the todos are remarks to _current_ code that I came along while implementing this patch. So they are references to code that should be improved in time, but not necessarily in the "uber" patch.