@bolkedebruin 2016-06-01
@plypaul @mistercrunch I'm assuming #1559 is referred to. That one does raise some questions with me and @jlowin, but I dont mind working with @plypaul to make it work together. It should not be too difficult as #1559 is basically wrapping the areas I touched mostly.