@plypaul 2016-06-01
Yeah, as @bolkedebruin mentioned, this should be mostly compatible with #1559. To clarify, what's the difference between the `SCHEDULED` and `QUEUED` states?