@criccomini 2016-06-02
@bolkedebruin so is the order SCHEDULED -> QUEUED?