@plypaul 2016-06-03
Can you elaborate on the race condition that the `SCHEDULED` state helps to resolve?