@plypaul 2016-06-03
With the `SCHEDULED` state, is the following situation possible?

1. Task gets created by the scheduler in the `NONE` state
2. Scheduler looks at it and puts it in the `SCHEDULED` state
3. Before the task is sent to the executor, the scheduler crashes / killed and restarted.
4. Task stays in `SCHEDULED` state and doesn't get executed?