@plypaul 2016-06-03
Since the code in `TaskInstance.run()` somewhat mitigates task instances getting queued more than once, what kind of issues have you been seeing prior to this PR?