@plypaul 2016-06-03
Prior to adding the `SCHEDULED` state, I take it that you were seeing task instances getting queued multiple times in the executor? Since the code in `TaskInstance.run()` seems to (somewhat) mitigate this by detecting the `RUNNING` state, what sort of issues were you seeing?