@bolkedebruin 2016-06-03
On the occurrence of the race condition please see the comment in the executor by @jlowin. As said when people get more complex and start running two local executors for example the changes for this to occur would be there for every scheduler loop so for it to occur at a certain point in time would be 99% certain. People might just not have seen it.