@plypaul 2016-06-08
The condition where task instances can be orphaned doesn't require `SIGKILL` - if the scheduler is stopped by `SIGTERM` after the task is sent to the executor but before the task is run, all those task instances will be left in the `SCHEDULED` state, right?

This could be a problem in our setup where there can be 1000's of task instances waiting for a slot in the executor to run. That queue can take hours to clear, and if the scheduler is restarted at any point in that window, we'll have a bunch of orphaned tasks.

We also use the `CeleryExecutor`, which relies on an external queue mechanism. The queue can be cleared (for example, the Redis instance dies or needs to be restarted), and when this happens, we'll also have a bunch of orphaned tasks.

Overall, it would be great for Airflow to be resilient to restarts and failures. Machines die and services restart all the time, and the oncall who wakes up needs to have simple remediation procedures for handling the failure. Without a way to recover from these cases, it would make the oncall's life much more difficult.